requirements.txt
dotenv
tkinter
tkcalendar
mysql.connector